+++ tests/test_pandas_primitives.py
@@ -15,7 +15,7 @@ except ImportError:
 from mplchart.chart import Chart  # noqa: E402
 from mplchart.samples import sample_prices  # noqa: E402
 from mplchart.primitives import (  # noqa: E402
-    Candlesticks, OHLC, Volume,
+    Candlesticks, HeikinAshi, OHLC, Volume,
     LinePlot, AreaPlot, BarPlot,
     Swings, ZigZag, Stripes, Markers,
     HLine, VLine, TrendLines,
@@ -27,6 +27,7 @@ FREQS = ["daily", "hourly", "minute"]
 
 PRIMITIVES = [
     Candlesticks(),
+    HeikinAshi(),
     OHLC(),
     LinePlot("close"),
     Volume(),